Shift store init to own directory
