+++ src/store/layout/reducers.js
@@ -16,11 +16,12 @@ const ui = (state = defaultState, action) => {
       return Object.assign({}, state, {
         userPreferencesModalOpen: action.state,
       });
-    case 'SET_LABELLING_FLOW_DATA':
+    case 'SET_LABELLING_FLOW_DATA': {
       const labelling = Object.assign({}, action.labellingFlowData);
 
       return Object.assign({}, state, { labelling });
-    case 'SET_TOOL_CONTEXT_MENU_DATA':
+    }
+    case 'SET_TOOL_CONTEXT_MENU_DATA': {
       const contextMenu = Object.assign({}, state.contextMenu);
 
       contextMenu[action.viewportIndex] = Object.assign(
@@ -29,6 +30,7 @@ const ui = (state = defaultState, action) => {
       );
 
       return Object.assign({}, state, { contextMenu });
+    }
     case 'RESET_LABELLING_AND_CONTEXT_MENU':
       return Object.assign({}, state, {
         labelling: defaultState.labelling,
